+++ test/HandleHumanLocalMUnitTests.hs
@@ -22,15 +22,12 @@ import           Game.LambdaHack.Common.State
 import           Game.LambdaHack.Content.TileKind
 import           Game.LambdaHack.Definition.DefsInternal
   (toContentId, toContentSymbol)
-import           Game.LambdaHack.Definition.Flavour
 
+import Game.LambdaHack.Common.Kind (emptyMultiGroupItem)
 import UnitTestHelpers
 
-stubItem :: Item
-stubItem = Item { jkind = IdentityObvious (toContentId 0), jfid = Nothing, jflavour = dummyFlavour }
-
 testItemFull :: ItemFull
-testItemFull = ItemFull { itemBase = stubItem, itemKindId = toContentId 0, itemKind = testItemKind, itemDisco = ItemDiscoFull emptyAspectRecord, itemSuspect = False }
+testItemFull = ItemFull { itemBase = stubItem, itemKindId = toContentId 0, itemKind = emptyMultiGroupItem, itemDisco = ItemDiscoFull emptyAspectRecord, itemSuspect = False }
 
 handleHumanLocalMUnitTests :: TestTree
 handleHumanLocalMUnitTests = testGroup "handleHumanLocalMUnitTests"